rest_api/requirements.txt
Django==3.2
djangorestframework
django-cors-headers
djoser
psycopg2-binary
python-dotenv